# apps/web/openapi/generated/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CourseLit REST API",
        "version": "1.0.0",
        "description": "OpenAPI documentation for the public CourseLit REST API surface."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current CourseLit school origin"
        }
    ],
...
    "paths": {
...
        "/api/products/{productId}": {
            "get": {
                "tags": ["Products"],
                "summary": "Get a product",
                "operationId": "getProduct",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["Products"],
                "summary": "Update product metadata",
                "operationId": "updateProduct",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            },
                            "example": {
                                "title": "AI Foundations",
                                "slug": "ai-foundations",
                                "description": "{\"type\":\"doc\",\"content\":[{\"type\":\"paragraph\",\"content\":[{\"type\":\"text\",\"text\":\"Updated course description.\"}]}]}",
                                "published": false,
                                "privacy": "unlisted",
                                "tags": ["ai", "beginner"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body, non-object body, or unsupported product field.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Product could not be updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Products"],
                "summary": "Delete a product",
                "operationId": "deleteProduct",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Product could not be deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PublicApiErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "required": ["code", "message"]
                    }
                },
                "required": ["error"]
            },
            "OkResponse": {
                "type": "object",
                "properties": {
                    "ok": {
                        "type": "boolean"
                    }
                }
            },
            "PaymentPlan": {
                "type": "object",
                "properties": {
                    "planId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": ["free", "onetime", "emi", "subscription"]
                    },
                    "entityId": {
                        "type": "string"
                    },
                    "entityType": {
                        "type": "string"
                    },
                    "oneTimeAmount": {
                        "type": "number"
                    },
                    "emiAmount": {
                        "type": "number"
                    },
                    "emiTotalInstallments": {
                        "type": "number"
                    },
                    "subscriptionMonthlyAmount": {
                        "type": "number"
                    },
                    "subscriptionYearlyAmount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "isDefault": {
                        "type": "boolean"
                    }
                }
            },
...
            "Product": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": ["course", "download", "blog"]
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "published": {
                        "type": "boolean"
                    },
                    "privacy": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "featuredImage": {
                        "type": "object"
                    },
                    "lecturerId": {
                        "type": "string",
                        "nullable": true
                    },
                    "moderatorId": {
                        "type": "string",
                        "nullable": true
                    },
                    "pageId": {
                        "type": "string"
                    },
                    "defaultPaymentPlan": {
                        "type": "string"
                    },
                    "paymentPlans": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaymentPlan"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ProductUpdateRequest": {
                "type": "object",
                "description": "Payload for updating product metadata. A product must have at least one payment plan before it can be published.",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Product title shown in CourseLit.",
                        "example": "AI Foundations"
                    },
                    "slug": {
                        "type": "string",
                        "description": "Optional URL slug.",
                        "example": "ai-foundations"
                    },
                    "description": {
                        "type": "string",
                        "description": "Optional product/blog description. Send this as a JSON-stringified Tiptap/ProseMirror document, for example `JSON.stringify({ type: \"doc\", content: [] })`. Do not use a `content` field on this endpoint.",
                        "example": "{\"type\":\"doc\",\"content\":[{\"type\":\"paragraph\",\"content\":[{\"type\":\"text\",\"text\":\"Updated course description.\"}]}]}"
                    },
                    "published": {
                        "type": "boolean",
                        "description": "Whether the product is published. Existing CourseLit publishing checks still apply.",
                        "example": false
                    },
                    "privacy": {
                        "type": "string",
                        "description": "Existing CourseLit product privacy value.",
                        "example": "unlisted"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "featuredImage": {
                        "type": "object"
                    },
                    "lecturerId": {
                        "type": "string",
                        "nullable": true
                    },
                    "moderatorId": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key",
                "description": "API key created in CourseLit dashboard settings. The legacy `apikey` request-body field is still accepted but deprecated."
            }
        }
    }
}